# ohmyfood/docs/openapi.json
{"openapi":"3.0.0","paths":{"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_getMe","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/api/users":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/users/me/addresses":{"get":{"operationId":"UsersController_getMyAddresses","parameters":[],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]},"post":{"operationId":"UsersController_createAddress","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressDto"}}}},"responses":{"201":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/users/me/addresses/{addressId}":{"put":{"operationId":"UsersController_updateAddress","parameters":[{"name":"addressId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAddressDto"}}}},"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_deleteAddress","parameters":[{"name":"addressId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/restaurants":{"get":{"operationId":"RestaurantsController_list","parameters":[{"name":"category","required":false,"in":"query","schema":{"type":"string"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"take","required":false,"in":"query","schema":{"type":"number"}},{"name":"skip","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["restaurants"]},"post":{"operationId":"RestaurantsController_create","parameters":[],"responses":{"201":{"description":""}},"tags":["restaurants"],"security":[{"bearer":[]}]}},"/api/restaurants/{id}":{"get":{"operationId":"RestaurantsController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["restaurants"]}},"/api/restaurants/{id}/stats":{"get":{"operationId":"RestaurantsController_getStats","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["restaurants"],"security":[{"bearer":[]}]}},"/api/restaurants/{id}/orders":{"get":{"operationId":"RestaurantsController_getOrders","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["restaurants"],"security":[{"bearer":[]}]}},"/api/restaurants/{restaurantId}/menu":{"get":{"operationId":"MenuController_list","parameters":[{"name":"restaurantId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["menu"]},"post":{"operationId":"MenuController_create","parameters":[{"name":"restaurantId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["menu"],"security":[{"bearer":[]}]}},"/api/restaurants/{restaurantId}/menu/{id}":{"get":{"operationId":"MenuController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["menu"]},"put":{"operationId":"MenuController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["menu"],"security":[{"bearer":[]}]},"delete":{"operationId":"MenuController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["menu"],"security":[{"bearer":[]}]}},"/api/restaurants/{restaurantId}/menu/{menuItemId}/option-groups":{"post":{"operationId":"MenuController_createOptionGroup","parameters":[{"name":"menuItemId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["menu"],"security":[{"bearer":[]}]}},"/api/restaurants/{restaurantId}/menu/option-groups/{optionGroupId}":{"put":{"operationId":"MenuController_updateOptionGroup","parameters":[{"name":"optionGroupId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["menu"],"security":[{"bearer":[]}]},"delete":{"operationId":"MenuController_deleteOptionGroup","parameters":[{"name":"optionGroupId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["menu"],"security":[{"bearer":[]}]}},"/api/restaurants/{restaurantId}/menu/option-groups/{optionGroupId}/options":{"post":{"operationId":"MenuController_createOption","parameters":[{"name":"optionGroupId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["menu"],"security":[{"bearer":[]}]}},"/api/restaurants/{restaurantId}/menu/options/{optionId}":{"put":{"operationId":"MenuController_updateOption","parameters":[{"name":"optionId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["menu"],"security":[{"bearer":[]}]},"delete":{"operationId":"MenuController_deleteOption","parameters":[{"name":"optionId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["menu"],"security":[{"bearer":[]}]}},"/api/orders/user/{userId}":{"get":{"operationId":"OrdersController_listForUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"],"security":[{"bearer":[]}]},"post":{"operationId":"OrdersController_create","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["orders"],"security":[{"bearer":[]}]}},"/api/orders/restaurant/{restaurantId}":{"get":{"operationId":"OrdersController_listForRestaurant","parameters":[{"name":"restaurantId","required":true,"in":"path","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"],"security":[{"bearer":[]}]}},"/api/orders/available/courier":{"get":{"operationId":"OrdersController_listAvailableForCourier","parameters":[{"name":"lat","required":false,"in":"query","schema":{"type":"number"}},{"name":"lng","required":false,"in":"query","schema":{"type":"number"}},{"name":"maxDistance","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["orders"],"security":[{"bearer":[]}]}},"/api/orders/{id}":{"get":{"operationId":"OrdersController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"],"security":[{"bearer":[]}]}},"/api/orders/{id}/status":{"put":{"operationId":"OrdersController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"],"security":[{"bearer":[]}]}},"/api/orders/{id}/assign-courier":{"put":{"operationId":"OrdersController_assignCourier","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"],"security":[{"bearer":[]}]}},"/payments/stripe/webhook":{"post":{"operationId":"PaymentsController_handleStripeWebhook","parameters":[{"name":"stripe-signature","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["payments"]}},"/api/promos":{"get":{"operationId":"PromosController_list","parameters":[],"responses":{"200":{"description":""}},"tags":["promos"]}},"/api/support/tickets":{"post":{"operationId":"SupportController_createTicket","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSupportTicketDto"}}}},"responses":{"201":{"description":""}},"tags":["support"],"security":[{"bearer":[]}]},"get":{"operationId":"SupportController_listTickets","parameters":[],"responses":{"200":{"description":""}},"tags":["support"],"security":[{"bearer":[]}]}},"/api/support/tickets/{id}":{"get":{"operationId":"SupportController_getTicket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["support"],"security":[{"bearer":[]}]},"put":{"operationId":"SupportController_updateTicket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSupportTicketDto"}}}},"responses":{"200":{"description":""}},"tags":["support"],"security":[{"bearer":[]}]}},"/api/admin/summary":{"get":{"operationId":"AdminController_summary","parameters":[],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/admin/live-orders":{"get":{"operationId":"AdminController_liveOrders","parameters":[],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/admin/restaurants":{"get":{"operationId":"AdminController_listRestaurants","parameters":[],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/admin/restaurants/{id}/approve":{"put":{"operationId":"AdminController_approveRestaurant","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/admin/restaurants/{id}/suspend":{"put":{"operationId":"AdminController_suspendRestaurant","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/admin/couriers":{"get":{"operationId":"AdminController_listCouriers","parameters":[],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/admin/couriers/{id}/approve":{"put":{"operationId":"AdminController_approveCourier","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/admin/couriers/{id}/suspend":{"put":{"operationId":"AdminController_suspendCourier","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/admin/orders":{"get":{"operationId":"AdminController_listOrders","parameters":[],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/admin/orders/{id}/cancel":{"put":{"operationId":"AdminController_cancelOrder","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/admin/orders/{id}/reassign-courier":{"put":{"operationId":"AdminController_reassignCourier","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}}},"info":{"title":"OhMyFood API","description":"MVP API para plataforma de entregas OhMyFood","version":"0.1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"novo@ohmyfood.pt"},"password":{"type":"string","minLength":6},"role":{"type":"string","enum":["CUSTOMER","COURIER","RESTAURANT","ADMIN"],"default":"CUSTOMER"},"displayName":{"type":"string"}},"required":["email","password","role"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"cliente@ohmyfood.pt"},"password":{"type":"string","minLength":6}},"required":["email","password"]},"CreateAddressDto":{"type":"object","properties":{"label":{"type":"string"},"street":{"type":"string"},"number":{"type":"string"},"complement":{"type":"string"},"postalCode":{"type":"string"},"city":{"type":"string"},"country":{"type":"string","default":"Portugal"},"lat":{"type":"number"},"lng":{"type":"number"},"instructions":{"type":"string"},"isDefault":{"type":"boolean","default":false}},"required":["label","street","number","postalCode","city"]},"UpdateAddressDto":{"type":"object","properties":{"label":{"type":"string"},"street":{"type":"string"},"number":{"type":"string"},"complement":{"type":"string"},"postalCode":{"type":"string"},"city":{"type":"string"},"country":{"type":"string"},"lat":{"type":"number"},"lng":{"type":"number"},"instructions":{"type":"string"},"isDefault":{"type":"boolean"}}},"CreateOrderItemDto":{"type":"object","properties":{"menuItemId":{"type":"string"},"quantity":{"type":"number","minimum":1},"addons":{"type":"array","items":{"type":"string"}}},"required":["menuItemId","quantity"]},"CreateOrderDto":{"type":"object","properties":{"restaurantId":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/CreateOrderItemDto"}},"itemsTotalCents":{"type":"number"},"deliveryFeeCents":{"type":"number"},"serviceFeeCents":{"type":"number"}},"required":["restaurantId","items","itemsTotalCents","deliveryFeeCents","serviceFeeCents"]},"CreateSupportTicketDto":{"type":"object","properties":{"type":{"type":"string"},"subject":{"type":"string"},"description":{"type":"string"},"orderId":{"type":"string"}},"required":["type","subject","description"]},"UpdateSupportTicketDto":{"type":"object","properties":{"status":{"type":"string"},"adminNotes":{"type":"string"}}}}}}